# CUDA: optimize MMQ int8 tensor core performance