This PR adds the following optimizations for the CUDA MMQ kernels using int8 tensor cores:

* After https://github.com/ggerganov/llama.cpp/pull/8018 the restrictions on tile sizes have become much looser so I changed the tile sizes for >= 48 tokens in such a way that less data needs to be loaded from shared memory per tensor core operation. This necessitates that the tile size in terms of parallel tokens is a multiple of 16 (instead of 8). As a side effect that also means that fewer kernel versions need to be compiled so compile time and binary size go down.
* I changed the shared memory layout for the quantized data in such a way that there are fewer shared memory bank conflicts when loading the data.
* I exposed the `ldmatrix` PTX instruction to load data in blocks of 16 bytes instead of 4.

<details>
<summary>Performance vs. master MMQ</summary>

| GPU        | Model             |     Microbatch size | Test     |     t/s master |     t/s cuda-mmq-2xa-3 |     Speedup |
| :--------- | :---------------- | ------------------: | :------- | -------------: | ---------------------: | ----------: |
| RTX 4090   | llama 8B Q2_K_M   |                  16 | pp2048   |        1959.31 |                2058.73 |        1.05 |
| RTX 4090   | llama 8B Q2_K_M   |                  32 | pp2048   |        3041.88 |                3093.97 |        1.02 |
| RTX 4090   | llama 8B Q2_K_M   |                  63 | pp2048   |        4327.05 |                4566.56 |        1.06 |
| RTX 4090   | llama 8B Q2_K_M   |                 128 | pp2048   |        5779.88 |                6088.58 |        1.05 |
| RTX 4090   | llama 8B Q2_K_M   |                 256 | pp2048   |        7145.78 |                7726.39 |        1.08 |
| RTX 4090   | llama 8B Q2_K_M   |                 512 | pp2048   |        7553.00 |                8194.04 |        1.08 |
| RTX 4090   | llama 8B Q2_K_M   |                1024 | pp2048   |        7458.42 |                8046.76 |        1.08 |
| RTX 4090   | llama 8B Q2_K_M   |                2048 | pp2048   |        6906.43 |                7424.69 |        1.08 |
| RTX 4090   | llama 8B Q3_K_S   |                  16 | pp2048   |        1777.22 |                1945.13 |        1.09 |
| RTX 4090   | llama 8B Q3_K_S   |                  32 | pp2048   |        3015.02 |                3090.91 |        1.03 |
| RTX 4090   | llama 8B Q3_K_S   |                  63 | pp2048   |        4620.00 |                4446.98 |        0.96 |
| RTX 4090   | llama 8B Q3_K_S   |                 128 | pp2048   |        6713.58 |                6592.28 |        0.98 |
| RTX 4090   | llama 8B Q3_K_S   |                 256 | pp2048   |        8489.05 |                8324.30 |        0.98 |
| RTX 4090   | llama 8B Q3_K_S   |                 512 | pp2048   |        8968.09 |                8768.22 |        0.98 |
| RTX 4090   | llama 8B Q3_K_S   |                1024 | pp2048   |        8819.05 |                8589.54 |        0.97 |
| RTX 4090   | llama 8B Q3_K_S   |                2048 | pp2048   |        8020.62 |                7882.24 |        0.98 |
| RTX 4090   | llama 8B Q4_0     |                  16 | pp2048   |        1970.65 |                1988.00 |        1.01 |
| RTX 4090   | llama 8B Q4_0     |                  32 | pp2048   |        3431.49 |                3445.99 |        1.00 |
| RTX 4090   | llama 8B Q4_0     |                  63 | pp2048   |        5357.19 |                5284.59 |        0.99 |
| RTX 4090   | llama 8B Q4_0     |                 128 | pp2048   |        7509.03 |                7502.92 |        1.00 |
| RTX 4090   | llama 8B Q4_0     |                 256 | pp2048   |        9609.65 |                9726.08 |        1.01 |
| RTX 4090   | llama 8B Q4_0     |                 512 | pp2048   |       10285.00 |               10683.60 |        1.04 |
| RTX 4090   | llama 8B Q4_0     |                1024 | pp2048   |       10018.49 |               10364.80 |        1.03 |
| RTX 4090   | llama 8B Q4_0     |                2048 | pp2048   |        9155.64 |                9417.97 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                  16 | pp2048   |        1862.71 |                1874.43 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                  32 | pp2048   |        2979.06 |                3018.03 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                  63 | pp2048   |        5145.03 |                5312.52 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                 128 | pp2048   |        7229.84 |                7462.47 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                 256 | pp2048   |        9130.95 |                9438.28 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                 512 | pp2048   |        9747.77 |               10059.40 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                1024 | pp2048   |        9545.98 |                9830.56 |        1.03 |
| RTX 4090   | llama 8B Q4_1     |                2048 | pp2048   |        8699.96 |                8906.72 |        1.02 |
| RTX 4090   | llama 8B Q4_K_S   |                  16 | pp2048   |        1982.75 |                1987.60 |        1.00 |
| RTX 4090   | llama 8B Q4_K_S   |                  32 | pp2048   |        3361.63 |                3500.66 |        1.04 |
| RTX 4090   | llama 8B Q4_K_S   |                  63 | pp2048   |        5149.95 |                5331.79 |        1.04 |
| RTX 4090   | llama 8B Q4_K_S   |                 128 | pp2048   |        6994.57 |                7288.84 |        1.04 |
| RTX 4090   | llama 8B Q4_K_S   |                 256 | pp2048   |        8788.43 |                9232.26 |        1.05 |
| RTX 4090   | llama 8B Q4_K_S   |                 512 | pp2048   |        9403.70 |                9877.78 |        1.05 |
| RTX 4090   | llama 8B Q4_K_S   |                1024 | pp2048   |        9294.39 |                9741.86 |        1.05 |
| RTX 4090   | llama 8B Q4_K_S   |                2048 | pp2048   |        8552.56 |                8897.37 |        1.04 |
| RTX 4090   | llama 8B Q5_0     |                  16 | pp2048   |        1660.02 |                1713.62 |        1.03 |
| RTX 4090   | llama 8B Q5_0     |                  32 | pp2048   |        3029.92 |                3073.33 |        1.01 |
| RTX 4090   | llama 8B Q5_0     |                  63 | pp2048   |        4691.46 |                5140.13 |        1.10 |
| RTX 4090   | llama 8B Q5_0     |                 128 | pp2048   |        6847.64 |                7470.47 |        1.09 |
| RTX 4090   | llama 8B Q5_0     |                 256 | pp2048   |        8927.18 |                9696.99 |        1.09 |
| RTX 4090   | llama 8B Q5_0     |                 512 | pp2048   |        9707.85 |               10536.44 |        1.09 |
| RTX 4090   | llama 8B Q5_0     |                1024 | pp2048   |        9529.31 |               10299.92 |        1.08 |
| RTX 4090   | llama 8B Q5_0     |                2048 | pp2048   |        8725.90 |                9363.21 |        1.07 |
| RTX 4090   | llama 8B Q5_1     |                  16 | pp2048   |        1572.10 |                1627.59 |        1.04 |
| RTX 4090   | llama 8B Q5_1     |                  32 | pp2048   |        2527.92 |                2727.62 |        1.08 |
| RTX 4090   | llama 8B Q5_1     |                  63 | pp2048   |        4273.50 |                4395.52 |        1.03 |
| RTX 4090   | llama 8B Q5_1     |                 128 | pp2048   |        6769.11 |                7198.66 |        1.06 |
| RTX 4090   | llama 8B Q5_1     |                 256 | pp2048   |        8493.96 |                9079.53 |        1.07 |
| RTX 4090   | llama 8B Q5_1     |                 512 | pp2048   |        9064.31 |                9827.46 |        1.08 |
| RTX 4090   | llama 8B Q5_1     |                1024 | pp2048   |        9008.68 |                9696.96 |        1.08 |
| RTX 4090   | llama 8B Q5_1     |                2048 | pp2048   |        8304.00 |                8897.12 |        1.07 |
| RTX 4090   | llama 8B Q5_K_S   |                  16 | pp2048   |        1690.31 |                1717.22 |        1.02 |
| RTX 4090   | llama 8B Q5_K_S   |                  32 | pp2048   |        2760.78 |                3021.09 |        1.09 |
| RTX 4090   | llama 8B Q5_K_S   |                  63 | pp2048   |        4415.75 |                4883.11 |        1.11 |
| RTX 4090   | llama 8B Q5_K_S   |                 128 | pp2048   |        6646.87 |                6905.05 |        1.04 |
| RTX 4090   | llama 8B Q5_K_S   |                 256 | pp2048   |        8316.00 |                8707.63 |        1.05 |
| RTX 4090   | llama 8B Q5_K_S   |                 512 | pp2048   |        8924.37 |                9348.09 |        1.05 |
| RTX 4090   | llama 8B Q5_K_S   |                1024 | pp2048   |        8858.57 |                9265.80 |        1.05 |
| RTX 4090   | llama 8B Q5_K_S   |                2048 | pp2048   |        8203.50 |                8525.00 |        1.04 |
| RTX 4090   | llama 8B Q6_K     |                  16 | pp2048   |        1445.48 |                1466.03 |        1.01 |
| RTX 4090   | llama 8B Q6_K     |                  32 | pp2048   |        2736.55 |                2774.79 |        1.01 |
| RTX 4090   | llama 8B Q6_K     |                  63 | pp2048   |        4473.27 |                4541.41 |        1.02 |
| RTX 4090   | llama 8B Q6_K     |                 128 | pp2048   |        6516.45 |                6720.65 |        1.03 |
| RTX 4090   | llama 8B Q6_K     |                 256 | pp2048   |        8328.57 |                8696.60 |        1.04 |
| RTX 4090   | llama 8B Q6_K     |                 512 | pp2048   |        9004.15 |                9461.48 |        1.05 |
| RTX 4090   | llama 8B Q6_K     |                1024 | pp2048   |        8816.76 |                9285.52 |        1.05 |
| RTX 4090   | llama 8B Q6_K     |                2048 | pp2048   |        8056.57 |                8451.71 |        1.05 |
| RTX 4090   | llama 8B Q8_0     |                  16 | pp2048   |        1271.09 |                1264.01 |        0.99 |
| RTX 4090   | llama 8B Q8_0     |                  32 | pp2048   |        2362.24 |                2369.64 |        1.00 |
| RTX 4090   | llama 8B Q8_0     |                  63 | pp2048   |        4111.23 |                4141.30 |        1.01 |
| RTX 4090   | llama 8B Q8_0     |                 128 | pp2048   |        6820.92 |                6870.90 |        1.01 |
| RTX 4090   | llama 8B Q8_0     |                 256 | pp2048   |        9500.61 |                9903.76 |        1.04 |
| RTX 4090   | llama 8B Q8_0     |                 512 | pp2048   |       10607.57 |               11209.17 |        1.06 |
| RTX 4090   | llama 8B Q8_0     |                1024 | pp2048   |       10481.10 |               11106.60 |        1.06 |
| RTX 4090   | llama 8B Q8_0     |                2048 | pp2048   |        9610.52 |               10098.05 |        1.05 |
| RTX 3090   | llama 8B Q2_K_M   |                  16 | pp2048   |        1048.92 |                1083.08 |        1.03 |
| RTX 3090   | llama 8B Q2_K_M   |                  32 | pp2048   |        1535.68 |                1531.48 |        1.00 |
| RTX 3090   | llama 8B Q2_K_M   |                  63 | pp2048   |        2053.18 |                2165.83 |        1.05 |
| RTX 3090   | llama 8B Q2_K_M   |                 128 | pp2048   |        2477.58 |                2625.68 |        1.06 |
| RTX 3090   | llama 8B Q2_K_M   |                 256 | pp2048   |        2753.83 |                2906.06 |        1.06 |
| RTX 3090   | llama 8B Q2_K_M   |                 512 | pp2048   |        2848.71 |                3033.38 |        1.06 |
| RTX 3090   | llama 8B Q2_K_M   |                1024 | pp2048   |        2885.83 |                3051.31 |        1.06 |
| RTX 3090   | llama 8B Q2_K_M   |                2048 | pp2048   |        2831.48 |                3008.83 |        1.06 |
| RTX 3090   | llama 8B Q3_K_S   |                  16 | pp2048   |         981.92 |                1109.08 |        1.13 |
| RTX 3090   | llama 8B Q3_K_S   |                  32 | pp2048   |        1583.54 |                1595.45 |        1.01 |
| RTX 3090   | llama 8B Q3_K_S   |                  63 | pp2048   |        2242.77 |                2119.10 |        0.94 |
| RTX 3090   | llama 8B Q3_K_S   |                 128 | pp2048   |        2957.15 |                2854.64 |        0.97 |
| RTX 3090   | llama 8B Q3_K_S   |                 256 | pp2048   |        3296.41 |                3161.27 |        0.96 |
| RTX 3090   | llama 8B Q3_K_S   |                 512 | pp2048   |        3400.63 |                3258.34 |        0.96 |
| RTX 3090   | llama 8B Q3_K_S   |                1024 | pp2048   |        3455.97 |                3309.41 |        0.96 |
| RTX 3090   | llama 8B Q3_K_S   |                2048 | pp2048   |        3374.90 |                3232.75 |        0.96 |
| RTX 3090   | llama 8B Q4_0     |                  16 | pp2048   |        1238.54 |                1301.00 |        1.05 |
| RTX 3090   | llama 8B Q4_0     |                  32 | pp2048   |        1929.33 |                1940.42 |        1.01 |
| RTX 3090   | llama 8B Q4_0     |                  63 | pp2048   |        2727.29 |                2722.48 |        1.00 |
| RTX 3090   | llama 8B Q4_0     |                 128 | pp2048   |        3469.76 |                3491.56 |        1.01 |
| RTX 3090   | llama 8B Q4_0     |                 256 | pp2048   |        3885.53 |                3924.80 |        1.01 |
| RTX 3090   | llama 8B Q4_0     |                 512 | pp2048   |        4030.46 |                4124.14 |        1.02 |
| RTX 3090   | llama 8B Q4_0     |                1024 | pp2048   |        4043.94 |                4138.45 |        1.02 |
| RTX 3090   | llama 8B Q4_0     |                2048 | pp2048   |        3911.72 |                4000.28 |        1.02 |
| RTX 3090   | llama 8B Q4_1     |                  16 | pp2048   |        1371.83 |                1409.82 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                  32 | pp2048   |        1794.42 |                1754.93 |        0.98 |
| RTX 3090   | llama 8B Q4_1     |                  63 | pp2048   |        2586.43 |                2676.11 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                 128 | pp2048   |        3271.37 |                3354.55 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                 256 | pp2048   |        3625.96 |                3734.81 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                 512 | pp2048   |        3763.37 |                3876.53 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                1024 | pp2048   |        3791.94 |                3884.89 |        1.02 |
| RTX 3090   | llama 8B Q4_1     |                2048 | pp2048   |        3650.00 |                3776.34 |        1.03 |
| RTX 3090   | llama 8B Q4_K_S   |                  16 | pp2048   |        1304.90 |                1343.82 |        1.03 |
| RTX 3090   | llama 8B Q4_K_S   |                  32 | pp2048   |        1820.00 |                1998.71 |        1.10 |
| RTX 3090   | llama 8B Q4_K_S   |                  63 | pp2048   |        2491.79 |                2656.05 |        1.07 |
| RTX 3090   | llama 8B Q4_K_S   |                 128 | pp2048   |        3096.96 |                3205.98 |        1.04 |
| RTX 3090   | llama 8B Q4_K_S   |                 256 | pp2048   |        3449.95 |                3588.17 |        1.04 |
| RTX 3090   | llama 8B Q4_K_S   |                 512 | pp2048   |        3600.86 |                3736.54 |        1.04 |
| RTX 3090   | llama 8B Q4_K_S   |                1024 | pp2048   |        3645.23 |                3780.41 |        1.04 |
| RTX 3090   | llama 8B Q4_K_S   |                2048 | pp2048   |        3568.97 |                3683.96 |        1.03 |
| RTX 3090   | llama 8B Q5_0     |                  16 | pp2048   |        1023.34 |                1094.85 |        1.07 |
| RTX 3090   | llama 8B Q5_0     |                  32 | pp2048   |        1747.94 |                1804.44 |        1.03 |
| RTX 3090   | llama 8B Q5_0     |                  63 | pp2048   |        2392.03 |                2744.89 |        1.15 |
| RTX 3090   | llama 8B Q5_0     |                 128 | pp2048   |        3233.48 |                3444.51 |        1.07 |
| RTX 3090   | llama 8B Q5_0     |                 256 | pp2048   |        3619.52 |                3870.99 |        1.07 |
| RTX 3090   | llama 8B Q5_0     |                 512 | pp2048   |        3765.05 |                4056.88 |        1.08 |
| RTX 3090   | llama 8B Q5_0     |                1024 | pp2048   |        3769.53 |                4098.97 |        1.09 |
| RTX 3090   | llama 8B Q5_0     |                2048 | pp2048   |        3664.35 |                3941.37 |        1.08 |
| RTX 3090   | llama 8B Q5_1     |                  16 | pp2048   |        1041.17 |                1125.51 |        1.08 |
| RTX 3090   | llama 8B Q5_1     |                  32 | pp2048   |        1508.22 |                1632.28 |        1.08 |
| RTX 3090   | llama 8B Q5_1     |                  63 | pp2048   |        2280.39 |                2443.51 |        1.07 |
| RTX 3090   | llama 8B Q5_1     |                 128 | pp2048   |        3048.77 |                3207.34 |        1.05 |
| RTX 3090   | llama 8B Q5_1     |                 256 | pp2048   |        3377.53 |                3569.53 |        1.06 |
| RTX 3090   | llama 8B Q5_1     |                 512 | pp2048   |        3517.54 |                3743.99 |        1.06 |
| RTX 3090   | llama 8B Q5_1     |                1024 | pp2048   |        3547.11 |                3794.01 |        1.07 |
| RTX 3090   | llama 8B Q5_1     |                2048 | pp2048   |        3431.20 |                3676.47 |        1.07 |
| RTX 3090   | llama 8B Q5_K_S   |                  16 | pp2048   |        1083.87 |                1143.45 |        1.05 |
| RTX 3090   | llama 8B Q5_K_S   |                  32 | pp2048   |        1562.63 |                1737.00 |        1.11 |
| RTX 3090   | llama 8B Q5_K_S   |                  63 | pp2048   |        2212.98 |                2400.75 |        1.08 |
| RTX 3090   | llama 8B Q5_K_S   |                 128 | pp2048   |        2934.14 |                3021.05 |        1.03 |
| RTX 3090   | llama 8B Q5_K_S   |                 256 | pp2048   |        3284.45 |                3375.86 |        1.03 |
| RTX 3090   | llama 8B Q5_K_S   |                 512 | pp2048   |        3412.32 |                3532.41 |        1.04 |
| RTX 3090   | llama 8B Q5_K_S   |                1024 | pp2048   |        3453.93 |                3553.91 |        1.03 |
| RTX 3090   | llama 8B Q5_K_S   |                2048 | pp2048   |        3387.06 |                3478.17 |        1.03 |
| RTX 3090   | llama 8B Q6_K     |                  16 | pp2048   |         991.17 |                1011.44 |        1.02 |
| RTX 3090   | llama 8B Q6_K     |                  32 | pp2048   |        1660.10 |                1715.38 |        1.03 |
| RTX 3090   | llama 8B Q6_K     |                  63 | pp2048   |        2385.54 |                2456.09 |        1.03 |
| RTX 3090   | llama 8B Q6_K     |                 128 | pp2048   |        2957.59 |                3075.06 |        1.04 |
| RTX 3090   | llama 8B Q6_K     |                 256 | pp2048   |        3289.25 |                3459.49 |        1.05 |
| RTX 3090   | llama 8B Q6_K     |                 512 | pp2048   |        3433.88 |                3596.13 |        1.05 |
| RTX 3090   | llama 8B Q6_K     |                1024 | pp2048   |        3467.94 |                3646.15 |        1.05 |
…